Make RangeVariable support floats and back all Remixer numbers by floats

+++ remixer_annotation/src/test/resources/inputs/RangeVariableMethodAnnotationTest/Correct.java
@@ -24,5 +24,5 @@ public class Correct {
    * Default value is implicitly 0, this has to move to 15 with minValue.
    */
   @RangeVariableMethod(minValue = 15, maxValue = 20, defaultValue = 17)
-  public void correct(Integer i){}
+  public void correct(Float i){}
 }
